Fixes static library build (-Duse_zig_module=false when running examples)
