+++ src/examples/video_play.zig
@@ -212,8 +212,10 @@ pub fn main() !void {
 
             // Log the supported formats
             const format: *pw.c.spa_pod = @ptrCast(@alignCast(pw.c.spa_pod_builder_pop(&b, &format_frame).?));
-            log.info("supported formats:", .{});
-            check(pw.c.spa_debugc_format(dbg_ctx, 2, null, format));
+            if (example_options.use_zig_module) {
+                log.info("supported formats:", .{});
+                check(pw.c.spa_debugc_format(dbg_ctx, 2, null, format));
+            }
 
             // Add the supported formats to our params
             params.appendBounded(format) catch @panic("OOB");
@@ -234,7 +236,9 @@ pub fn main() !void {
             check(pw.c.spa_pod_builder_id(&b, pw.c.SPA_VIDEO_FORMAT_DSP_F32));
 
             const format: *const pw.c.spa_pod = @ptrCast(@alignCast(pw.c.spa_pod_builder_pop(&b, &format_frame)));
-            check(pw.c.spa_debugc_format(dbg_ctx, 2, null, format));
+            if (example_options.use_zig_module) {
+                check(pw.c.spa_debugc_format(dbg_ctx, 2, null, format));
+            }
             params.appendBounded(format) catch @panic("OOB");
         }
 
@@ -389,8 +393,10 @@ fn onStreamParamChanged(userdata: ?*anyopaque, id: u32, param: [*c]const pw.c.sp
     if (param == null or id != pw.c.SPA_PARAM_Format) return;
 
     // Log the new format
-    log.info("got format:", .{});
-    check(pw.c.spa_debugc_format(dbg_ctx, 2, null, param));
+    if (example_options.use_zig_module) {
+        log.info("got format:", .{});
+        check(pw.c.spa_debugc_format(dbg_ctx, 2, null, param));
+    }
 
     // Parse the new format and reset our timer to the new interval
     var parsed: pw.c.spa_video_info_raw = undefined;
